ListarCompras

+++ Seciv/SECIV/CapaLogica/LogicaNegocio/Compra.cs
@@ -10,7 +10,7 @@ namespace CapaLogica.LogicaNegocio
 {
     public class Compra
     {
-        public Compra(string id, string cod_Compra, float monto_Compra, string fecha_Compra)
+        public Compra(string id, string cod_Compra, Double monto_Compra, string fecha_Compra)
         {
             this.id = id;
             Cod_Compra = cod_Compra;
@@ -18,7 +18,7 @@ namespace CapaLogica.LogicaNegocio
             Fecha_Compra = fecha_Compra;
         }
 
-        public Compra(string cod_Compra, float monto_Compra, string fecha_Compra)
+        public Compra(string cod_Compra, Double monto_Compra, string fecha_Compra)
         {
             Cod_Compra = cod_Compra;
             Monto_Compra = monto_Compra;
@@ -33,7 +33,7 @@ namespace CapaLogica.LogicaNegocio
         public string Cod_Compra { get; set; }
 
         [BsonElement("Monto_Compra")]
-        public float Monto_Compra { get; set; }
+        public Double Monto_Compra { get; set; }
 
         [BsonElement("Fecha_Compra")]
         public string Fecha_Compra { get; set; }
